+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [1.0.9] - 2025-02-26
+### Fixed
+- [Local Service] Caretaker not undo first memento state
+
+### Modified
+- [Local Service] Improved Caretaker TryUndo and TryRedo code.
+
 ## [1.0.8] - 2025-02-24
 ### Added
 - [Global Service] Add more states to GlobalMementoState